requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
Pillow
pypdf
pdf2image
pikepdf
python-docx
pytesseract